+++ lib/benchee/output/benchmark_printer.ex
@@ -79,6 +79,9 @@ defmodule Benchee.Output.BenchmarkPrinter do
   def fast_warning do
     IO.puts """
     Warning: The function you are trying to benchmark is super fast, making measures more unreliable! See: https://github.com/PragTob/benchee/wiki/Benchee-Warnings#fast-execution-warning
+
+    You may disable this warning by passing print: [fast_warning: false] as
+    configuration options.
     """
   end