+++ test/benchee_test.exs
@@ -111,7 +111,7 @@ defmodule BencheeTest do
 
     warnings = output
                |> String.split("\n")
-               |> Enum.filter(fn(line) -> line =~ ~r/fast/ end)
+               |> Enum.filter(fn(line) -> line =~ ~r/Warning.+fast/ end)
 
     assert Enum.count(warnings) == 1
   end